fix EventNumber and RunNumber type
git-svn-id: svn+ssh://svn.cern.ch/reps/atlasphys-hsg8/Physics/Higgs/HSG8/AnalysisCode/multileptons/xAOD/ttHMultiAna/trunk@238089 883ba7d9-fdd0-4202-9341-49aa55999ad8

+++ ttHMultilepton/ttHMultileptonLooseEventSaver.h
@@ -153,8 +153,8 @@ class ttHMultileptonLooseEventSaver : public top::EventSaverFlatNtuple {
   double m_weight_bTagSF_77_extrapolation_from_charm_down;
 
   //event info
-  unsigned int m_eventNumber;
-  unsigned int m_runNumber;
+  unsigned long long m_eventNumber;
+  uint32_t m_runNumber;
   unsigned int m_mcChannelNumber;
   float m_mu;
   float m_mu_unc;
